# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SAKBFANTASY",
        "description": "Api SAKBFANTASY Documentation",
        "contact": {
            "email": "info@sakb.com"
        },
        "license": {
            "name": "Sakb",
            "url": "http://sakb-co.com.sa/"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/player_master": {
            "post": {
                "tags": [
                    "game"
                ],
                "summary": "get all players of your game",
                "operationId": "player_master",
                "parameters": [
                    {
                        "name": "access-token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type-dev",
                        "in": "header",
                        "description": "ios or android",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ios"
                        }
                    },
                    {
                        "name": "val-dev",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "sakb"
                        }
                    },
                    {
                        "name": "lang",
                        "in": "header",
                        "description": "default lang is ar ( ar , en)",
                        "schema": {
                            "type": "string",
                            "default": "ar"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "error"
                    }
                }
            }
        },
...
        "/api/v1/player": {
            "post": {
                "tags": [
                    "game"
                ],
                "summary": "get data of player",
                "operationId": "player",
                "parameters": [
                    {
                        "name": "access-token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type-dev",
                        "in": "header",
                        "description": "ios or android",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ios"
                        }
                    },
                    {
                        "name": "val-dev",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "sakb"
                        }
                    },
                    {
                        "name": "lang",
                        "in": "header",
                        "description": "default lang is ar ( ar , en)",
                        "schema": {
                            "type": "string",
                            "default": "ar"
                        }
                    },
                    {
                        "name": "player_link",
                        "in": "query",
                        "description": "link of player",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "player_link"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/v1/add_player": {
            "post": {
                "tags": [
                    "game"
                ],
                "summary": "add player to game",
                "operationId": "add_player",
                "parameters": [
                    {
                        "name": "access-token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type-dev",
                        "in": "header",
                        "description": "ios or android",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ios"
                        }
                    },
                    {
                        "name": "val-dev",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "sakb"
                        }
                    },
                    {
                        "name": "lang",
                        "in": "header",
                        "description": "default lang is ar ( ar , en)",
                        "schema": {
                            "type": "string",
                            "default": "ar"
                        }
                    },
                    {
                        "name": "player_link",
                        "in": "query",
                        "description": "link of player",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "player_link"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success Note:add_player= 1 ok add ,0 not add,-1 found ,-2 cost not enough,-3 complete number 15 players,-6 finish num from this team,-7 finish num from this location"
                    },
                    "400": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/v1/delete_player": {
            "post": {
                "tags": [
                    "game"
                ],
                "summary": "delete player from game and also change player with another",
                "operationId": "delete_player",
                "parameters": [
                    {
                        "name": "access-token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type-dev",
                        "in": "header",
                        "description": "ios or android",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ios"
                        }
                    },
                    {
                        "name": "val-dev",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "sakb"
                        }
                    },
                    {
                        "name": "lang",
                        "in": "header",
                        "description": "default lang is ar ( ar , en)",
                        "schema": {
                            "type": "string",
                            "default": "ar"
                        }
                    },
                    {
                        "name": "eldwry_link",
                        "in": "query",
                        "description": "link of eldwry",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "eldwry_link"
                        }
                    },
                    {
                        "name": "player_link",
                        "in": "query",
                        "description": "link of player",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "player_link"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "error"
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}